loadEmployeeById, loadCustomerById, executeSingleSelect

+++ src/main/java/pl/coderslab/model/Employee.java
@@ -3,24 +3,24 @@ package pl.coderslab.model;
 import java.math.BigDecimal;
 
 public class Employee extends Person {
-    private BigDecimal hourly;
+    private double hourly;
     private String note;
 
     public Employee() {
     }
 
-    public Employee(int id, String name, String surname, String address, String phone, String email, String birthDate, BigDecimal hourly, String note) {
+    public Employee(int id, String name, String surname, String address, String phone, String email, String birthDate, double hourly, String note) {
         super(id, name, surname, address, phone, email, birthDate);
         setHourly(hourly);
         setNote(note);
     }
 
 
-    public BigDecimal getHourly() {
+    public double getHourly() {
         return hourly;
     }
 
-    public void setHourly(BigDecimal hourly) {
+    public void setHourly(double hourly) {
         this.hourly = hourly;
     }
